\name{RcppTestB-package}
\alias{RcppTestB-package}
\alias{RcppTestB}
\docType{package}
\title{
Dummy package part of Rcpp unit tests
}
\description{
Dummy package part of Rcpp unit tests
}
\details{
\tabular{ll}{
Package: \tab RcppTestB\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-04-10\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
